Create usepagination file and refactor search
